*Feat* Novo design para tela de séries do exercício

+++ app/src/main/java/br/com/myfitt/common/domain/SerieExercicio.kt
@@ -13,7 +13,7 @@ data class SerieExercicio(
     val dhFimDescanso: LocalDateTime?,
     val duracaoSegundos: Int,
     val segundosDescanso: Int,
-    val pesoKg: Int,
+    val pesoKg: Float,
     val repeticoes: Int,
     val finalizado: Boolean
 ){
